Version 2 using API and cosmosdb
