+++ .env
@@ -1,0 +1,37 @@
+VUE_APP_COSMOSDB_KEY=9AVSNKSKGQDHlilnc8OnGoJ0GLk8VMAyvM6S23WBAhrGhZ5Xo7zeJLRNIZX257ihAE8W6YiqnlWYACDbYS5Hsg==
+VUE_APP_FIREBASE_APIKEY=AIzaSyCvTULYKq_MtDkNqGAo2NZ1G4AO6GocjPI
+VUE_APP_FIREBASE_AUTHDOMAIN=jiracloudversion.firebaseapp.com
+VUE_APP_FIREBASE_PROJECTID=jiracloudversion
+VUE_APP_FIREBASE_STORAGEBUCKET=jiracloudversion.appspot.com
+VUE_APP_FIREBASE_MESSAGINGSENDERID=658437942329
+VUE_APP_FIREBASE_APPID=1:658437942329:web:b1b51814f8a9aa85e1b1e8
+VUE_APP_MEASUREMENTID=G-9H9DYG041P
+VUE_APP_OAUTH2_CLIENT_ID=658437942329-9sleksnppkmfcajuhhnk9e5mic547b6r.apps.googleusercontent.com
+VUE_APP_MV3_KEY="-----BEGIN PRIVATE KEY-----
+MIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQDTlXebHyH8zFyz
+zolP3lg9QTkucSZQNpfbwssE2gQjekh0EmqU9ad6zkoP7kR4Dysa3OBUaCAJHyNS
+2ucY1krkgVTrj4Kj+xCdr4fhT1NYyD94J/v2EE8yNZNDleS/U5f4j5TbYeLvM5ho
+TOKGSIxCMVOOnHiiLIQV8+uz9NzqN0hclhQCbRzeZtu0qan0i+RP17WWhfg/ESqn
+M6vqoA0nNH3ur91XtpJiGza//LWR10s1HiQNbbtxFyYeVIIx9D73zab6nld2YbW0
+YWg3jBelOLk4o9iwGUYG2bt8WOfh3Qyjt2M0ysM6YVuGa6I8qYCDP1wH0v0AIl9b
+BHbcnjhbAgMBAAECggEAXGU5/OCEjAuGM/LuXZVf/SfdW52HuKGOgSY4V98IzaFZ
+wG5GYtQ/lc4942UOO0uMbEemItXxn2ocl9MF9f+mZtDfj5p01z0NDX8YnFhysiOI
+K+VyDwsiOm/Q5kOvx6kuhRnWTFMbsY9bb7ibdz55GXLG5VEJEPDVoWyKMmSskA5W
+4tepBHCncRNsTDB2LSwGCJ+GTVBeSuwMC5GSPt2IHIEdGXDIV7Ur8nAolLn3ct0W
+60iQroa+tTDwQJdS9Q5RZ/z3H6ZS7WMRqUd7wQvQBmp+ie3hvGWZNybBR1i/YxfM
+ZBdMRNUjVu+cV+uge0F/5m3MhDh/14eimj95YJJ8sQKBgQDqRRryi1UBv8EonIUh
+Kv9aVTNWHBM8kc/N9Isfx1m2v206UC5TXYe7uG54MA+VubE7FLTAoZHayEivgh92
+7OBKYZHvh4gULVgI7dGLSw0t41SmcdXlVsptJv6nin83XajZWNnJEpB8W5yGMpiK
+M5/NJLbpm04WRFcQfC20GS8NQwKBgQDnNapzm3M0NpFUB3ifvmSESME0JlA+ioS4
+E1l3I8TLcV2SgpJN1b60rMTTDnV20bHHT2MVFpTVXL/aHF4WlP22uEHTOkE0AC4Q
+K6iD8S2MO/gVffqG7DgsvvlMsQr2J87pHJYgoZNS5pQSL2jJYczj82ZwK2hWyeV2
+X4N57EmrCQKBgQCCZvUZDXXEA3kBoh541/ZL0iFBTV9n18JUYksO5WiK0QN78iXX
+QbjMRV04iYgjT7nXqS70CvulTNco2sWo7THsNWOGcWrN0G+wiLCbWKt9X40/GR53
+ClijaZFWsUEfEINseIjB5BOJJXV7Pa6rPxE0n7sxTcGpo2DTTEwkc+A9AQKBgQDm
+3vaqY+IglBCACMnRMUurr/6EdXGRUb3/FQsz5VqgDz5xisJOyI2aWh5b5cczc2aC
+7+mvaBIMD4ZfZA4rnHDQIKNSKYxsQXW9neqyfYSO0F8EDvy3926Md9H2W57nd46r
+xWi/grfagvX+Rqrq1CbExApZHW8dopLnpa5ic6iICQKBgQDZLXtEYKAaLspj0Qx4
+A58d8UUUiaaWp/r2ErdJInlmDC16ZtUdHTzmIsD3hkfKdRLRtQ74zXSUQujsFoa5
+rN6/NHVHg46wifZ2LX5U89GYo1RAqEgqipbrelhY7dui57EJavW6olS3wHSk9E+j
+GGmjkHDYfkLSHCHgZXJnypuydQ==
+-----END PRIVATE KEY-----"
