+++ .gitignore
@@ -1,10 +1,28 @@
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
+
 node_modules
-dist
-extension/manifest.json
-extension/dist
 .DS_Store
-.env
+dist
+dist-ssr
+coverage
+*.local
 
-*.pem
+/cypress/videos/
+/cypress/screenshots/
 
-*.crx
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
