+++ package.json
@@ -1,48 +1,30 @@
 {
-  "name": "jira-cloud-versioner",
-  "version": "1.0.0",
-  "type": "module",
+  "name": "jira-versioner",
+  "version": "0.0.0",
+  "private": true,
   "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w"
-  },
-  "dependencies": {
-    "dotenv": "^16.0.0",
-    "rollup-plugin-alias": "^2.2.0",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-import-css": "^3.0.3",
-    "toastify-js": "^1.12.0",
-    "vue": "^3.0.4",
-    "webextension-polyfill": "^0.7.0",
-    "webextension-polyfill-ts": "^0.22.0"
+    "dev": "vite",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false"
   },
   "devDependencies": {
-    "@rollup/plugin-alias": "^3.1.1",
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^11.0.1",
-    "@rollup/plugin-replace": "^2.3.4",
-    "@types/chrome": "0.0.164",
-    "@types/fs-extra": "^9.0.13",
-    "@types/node": "^16.11.10",
-    "@vitejs/plugin-vue": "^1.9.3",
-    "esno": "^0.12.1",
-    "firebase": "^9.5.0",
-    "fs-extra": "^10.0.0",
-    "node-sass": "^8.0.0",
-    "npm-run-all": "^4.1.5",
-    "pinia": "^2.0.28",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.38.5",
-    "rollup-plugin-chrome-extension": "^3.6.10",
-    "rollup-plugin-empty-dir": "^1.0.4",
-    "rollup-plugin-inject-process-env": "^1.3.1",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-typescript2": "^0.31.0",
-    "rollup-plugin-vue": "^6.0.0",
-    "rollup-plugin-zip": "^1.0.1",
-    "sass-loader": "^13.2.0",
-    "typescript": "^4.4.3",
+    "@tsconfig/node18": "^18.2.2",
+    "@types/node": "^18.18.6",
+    "@vitejs/plugin-vue": "^4.4.0",
+    "@vue/tsconfig": "^0.4.0",
+    "firebase": "^10.5.0",
+    "npm-run-all2": "^6.1.1",
+    "toastify-js": "^1.12.0",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.11",
+    "vite-plugin-mkcert": "^1.16.0",
+    "vue": "^3.3.7",
+    "vue-tsc": "^1.8.19"
+  },
+  "dependencies": {
+    "@azure/cosmos": "^4.0.0",
     "vuedraggable": "^4.1.0"
   }
 }
