+++ src/CodeBase.ts
@@ -4,11 +4,11 @@ export class CodeBase {
     Name: string;
     Issues = new Array<JiraTicket>();
 
-    constructor(name) {
+    constructor(name: string) {
       this.Name = name;      
     }
 
-    AddIssue(issue) {
+    AddIssue(issue: any) {
       if (!this.Issues.find( i => i.Number == issue.Number)) {
         this.Issues.push(issue);
       }
