--- src/manifest.json
@@ -1,39 +1,0 @@
-{
-  "name": "Jira Cloud Versioner",
-  "version": "1.1",
-  "manifest_version": 3,
-  "action": {
-    "default_icon": {
-      "16": "assets/icon16.png",
-      "48": "assets/icon48.png",
-      "128": "assets/icon128.png"
-    },
-    "default_title": "Popup Sample",
-    "default_popup": "popup/index.html"
-  },
-  "icons": {
-    "16": "assets/icon16.png",
-    "48": "assets/icon48.png",
-    "128": "assets/icon128.png"
-  },
-  "background": {
-    "service_worker": "background.ts"
-  },
-  "content_scripts": [
-    {
-      "matches": ["https://dealeron.atlassian.net/*"],
-      "all_frames": true,            
-      "run_at": "document_start",
-      "js": ["content-scripts/main.ts"]
-    }
-  ],
-   "web_accessible_resources": [
-    {
-      "resources": [ "assets/*.png", "content-scripts/main.ts", "pinia-*.js" ],
-      "matches": [ "https://dealeron.atlassian.net/*" ]
-    }
-  ],  
-  "permissions": [
-    "identity", "storage", "tabs", "scripting"
-  ]
-}
