--- src/popup/Popup.vue
@@ -1,16 +1,0 @@
-<template>
-  <div>
-    <ChromeExImg :filename="'icon48.png'" />
-    <h3>Jira Cloud Versioner</h3>   
-  </div>
-</template>
-
-<script setup lang="ts">
-import { onMounted } from "vue";
-
-onMounted(() => {  
-});
-</script>
-
-<style scoped>
-</style>
