--- src/popup/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-<head>
-  <meta charset="UTF-8">
-  <title>Popup</title>
-</head>
-<body style="min-width: 200px;">
-  <div id="jcvApp"></div>
-  <script type="module" src="./main.ts"></script>
-</body>
-</html>
