--- src/types/css.ts
@@ -1,2 +1,0 @@
-export type BoxSizeUnit = 'px' | 'pt' | 'pc' | 'in' | 'Q' | 'mm' | 'cm' | '%';
-export type BoxSizeTarget = 'width' | 'max-width' | 'height' | 'max-height';
