--- src/utils/cssUtils.ts
@@ -1,5 +1,0 @@
-import { BoxSizeTarget, BoxSizeUnit } from '../types/css'
-
-export function generateBoxSize (size: number, boxSizeUnit: BoxSizeUnit, boxSizeTarget: BoxSizeTarget) {
-  return `${boxSizeTarget}: ${size}${boxSizeUnit};`
-}
